refactor: waitgroup tests into multiple files for better organization
- Deleted the original waitgroup.go file containing all test cases.

- Created waitgroup_basic_cases.go for basic Add/Done patterns.

- Created waitgroup_callbacks_cases.go for function passing patterns.

- Created waitgroup_extended_cases.go for mixed scenarios and edge cases.

- Created waitgroup_goroutine_cases.go for loop and goroutine patterns.
