+++ pkg/analyzer/testdata/src/waitgroup/waitgroup_basic_cases.go
@@ -1,0 +1,139 @@
+package waitgroup
+
+import "sync"
+
+// ========== WAITGROUP TESTS ==========
+
+// ---------- Basic Add/Done Patterns ----------
+
+// Add and Done inside a goroutine
+func GoodBasicAddDone() {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+	}()
+	wg.Wait()
+}
+
+// Correct WaitGroup usage with short declaration
+func GoodWaitGroupShortDecl() {
+	wg := sync.WaitGroup{} // short declaration
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+	}()
+	wg.Wait()
+}
+
+// Add and Done in separate functions
+func GoodFuncAddDone() {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go handleExternalWork(&wg)
+	wg.Wait()
+}
+
+// Add and Wait with a gouroutine doing Done()
+func GoodAddBeforeWait() {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+	}()
+	wg.Wait()
+}
+
+// No Add, no Done (legal, Wait returns immediately)
+func GoodNoAddNoDone() {
+	var wg sync.WaitGroup
+	wg.Wait() // returns immediately
+}
+
+// Add without Done (counter never decremented)
+func BadAddWithoutDone() {
+	var wg sync.WaitGroup
+	wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+	wg.Wait()
+}
+
+// Bad Waitgroup with short declaration
+func BadWaitGroupShortDecl() {
+	wg := sync.WaitGroup{} // short declaration
+	wg.Add(1)              // want "waitgroup 'wg' has Add without corresponding Done"
+	wg.Wait()
+}
+
+// Multiple Add, but only one Done (the rest remain pending)
+func BadMultipleAddOneDone() {
+	var wg sync.WaitGroup
+	wg.Add(2) // want "waitgroup 'wg' has Add without corresponding Done"
+	wg.Add(1)
+	wg.Done()
+	wg.Wait()
+}
+
+// Add without Done in a goroutine that never runs
+func BadExtraDone() {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	wg.Done()
+	wg.Done() // want "waitgroup 'wg' has Done without corresponding Add"
+	wg.Wait()
+}
+
+// ---------- Wait Ordering Patterns ----------
+
+// Add after Wait (illegal, Wait should be called after all Adds)
+func BadAddAfterWait() {
+	var wg sync.WaitGroup
+	wg.Wait()
+	go func() {
+		wg.Add(1) // want "waitgroup 'wg' Add called after Wait"
+		wg.Done()
+	}()
+}
+
+// Edge case where Add is called after Wait, but in a different flow
+func EdgeCaseAddAfterWaitMainFlow() {
+	var wg sync.WaitGroup
+
+	wg.Wait()
+	wg.Add(1) // want "waitgroup 'wg' Add called after Wait"
+	wg.Done()
+}
+
+// Edge case where Wait is called without any Adds
+func EdgeCaseNoAddNoDoneNoGoroutine() {
+	var wg sync.WaitGroup
+	wg.Wait() // legal, returns immediately
+}
+
+// Edge case where Wait is called multiple times
+func EdgeCaseMultipleWaits() {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+	}()
+	wg.Wait()
+	wg.Wait()
+}
+
+// WaitGroup.Go handles Add/Done internally and should be accepted.
+func GoodWaitGroupGo() {
+	var wg sync.WaitGroup
+	wg.Go(func() {
+		doSomething()
+	})
+	wg.Wait()
+}
+
+// WaitGroup.Go after an empty Wait should be rejected, same as Add after Wait.
+func BadWaitGroupGoAfterWait() {
+	var wg sync.WaitGroup
+	wg.Wait()
+	wg.Go(func() { // want "waitgroup 'wg' Go called after Wait"
+		doSomething()
+	})
+}
