+++ pkg/analyzer/testdata/src/waitgroup/waitgroup_goroutine_cases.go
@@ -1,0 +1,206 @@
+package waitgroup
+
+import "sync"
+
+// ---------- Loop Patterns ----------
+
+// Add and Done inside a loop (typical worker pattern)
+func GoodLoopAddDone() {
+	var wg sync.WaitGroup
+	for i := 0; i < 3; i++ {
+		wg.Add(1)
+		go func() {
+			defer wg.Done()
+		}()
+	}
+	wg.Wait()
+}
+
+// Add inside a loop but Done may be missing in some paths
+func BadLoopAddMissingDone() {
+	var wg sync.WaitGroup
+	for i := 0; i < 2; i++ {
+		wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+		if i == 0 {
+			go func() {
+				wg.Done()
+			}()
+		}
+	}
+	wg.Wait()
+}
+
+// ---------- Goroutine Patterns ----------
+
+// Multiple goroutines with defer Done (should NOT trigger error)
+func GoodMultipleGoroutinesWithDeferDone() {
+	var wg sync.WaitGroup
+	wg.Add(2)
+	var errOrderConsumer, errReturnConsumer any
+	go func() {
+		defer wg.Done()
+		errOrderConsumer = doSomething()
+	}()
+	go func() {
+		defer wg.Done()
+		errReturnConsumer = doSomething()
+	}()
+	wg.Wait()
+
+	_ = errOrderConsumer
+	_ = errReturnConsumer
+}
+
+func GoodSwitchWithDefault() {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		x := 1
+		switch x {
+		case 2:
+			wg.Done()
+		default:
+			wg.Done()
+		}
+	}()
+	wg.Wait()
+}
+
+func GoodUnconditionalDone() {
+	var wg sync.WaitGroup
+	condition := false
+	wg.Add(1)
+	go func() {
+		if condition {
+			// some work
+		}
+		wg.Done() // This is outside the if, so it's unconditional
+	}()
+	wg.Wait()
+}
+
+// Add in a goroutine that never calls Done (e.g., due to deadlock or channel never sent)
+func BadAddNeverDone() {
+	var wg sync.WaitGroup
+	ch := make(chan struct{})
+	wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+	go func() {
+		<-ch // never sends, so Done is never called
+		wg.Done()
+	}()
+	wg.Wait()
+}
+
+// Add without Done in a goroutine that returns prematurely
+func BadAddDonePrematureReturn() {
+	var wg sync.WaitGroup
+	wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+	go func() {
+		return // forgot to call Done!
+		wg.Done()
+	}()
+	wg.Wait()
+}
+
+// Add without Done in a goroutine that panics
+func BadPanicWithoutRecover() {
+	var wg sync.WaitGroup
+	wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+	go func() {
+		panic("error") // Done is never called
+		wg.Done()
+	}()
+	wg.Wait()
+}
+
+// Add without Done in a goroutine with a conditional return
+func BadDeferWithConditionalReturn() {
+	var wg sync.WaitGroup
+	wg.Add(2) // want "waitgroup 'wg' has Add without corresponding Done"
+
+	go func() {
+		defer wg.Done()
+	}()
+
+	go func() {
+		if true {
+			return
+		}
+		defer wg.Done()
+	}()
+
+	wg.Wait()
+}
+
+// Add without Done in a goroutine with conditional Done
+func BadConditionalDone() {
+	var wg sync.WaitGroup
+	condition := false
+	wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+	go func() {
+		if condition { // condition is false, so Done is never called
+			wg.Done()
+		}
+	}()
+	wg.Wait()
+}
+
+// Also test with a more complex conditional
+func BadConditionalDoneComplex() {
+	var wg sync.WaitGroup
+	wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+	go func() {
+		x := 1
+		if x > 5 { // This is always false
+			wg.Done()
+		}
+	}()
+	wg.Wait()
+}
+
+// Test with switch statement
+func BadConditionalDoneSwitch() {
+	var wg sync.WaitGroup
+	wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+	go func() {
+		x := 1
+		switch x {
+		case 2: // x is 1, so this case won't match
+			wg.Done()
+		}
+	}()
+	wg.Wait()
+}
+
+func BadSwitchNoDefault() {
+	var wg sync.WaitGroup
+	wg.Add(1) // want "waitgroup 'wg' has Add without corresponding Done"
+	go func() {
+		x := 1
+		switch x {
+		case 2:
+			wg.Done()
+		case 3:
+			wg.Done()
+			// No default, and x=1 doesn't match any case
+		}
+	}()
+	wg.Wait()
+}
+
+// ---------- Panic Recovery Patterns ----------
+
+// Add/Done with panic recovery
+func GoodAddDoneWithPanicRecovery() {
+	var wg sync.WaitGroup
+	wg.Add(1)
+	go func() {
+		defer func() {
+			if r := recover(); r != nil {
+				wg.Done()
+			}
+		}()
+		panic("fail")
+	}()
+	wg.Wait()
+}
